requirements.txt
flask
pymupdf
pdfplumber
pandas
python-docx
openpyxl